apps/dashboard_service/requirements.txt
click
fastapi
httpx
pydantic
pydantic-settings
ruamel.yaml
uvicorn[standard]
SQLAlchemy
